+++ cola/widgets/dag.py
@@ -108,15 +108,19 @@ class ViewerMixin:
         """Return the currently selected commit object IDs"""
         return [i.commit for i in self.selected_items()]
 
-    def clicked_oid(self):
+    def clicked_oid(self, filtered=True):
         """Return the clicked or selected commit object ID"""
         if self.clicked:
-            return self.clicked.oid
-        return self.selected_oid()
+            oid = self.clicked.oid
+        else:
+            oid = self.selected_oid()
+        if filtered and oid and oid in (dag.STAGE, dag.WORKTREE):
+            oid = None
+        return oid
 
-    def with_oid(self, func):
+    def with_oid(self, func, filtered=True):
         """Run an operation with a commit object ID"""
-        oid = self.clicked_oid()
+        oid = self.clicked_oid(filtered=filtered)
         if oid:
             result = func(oid)
         else:
@@ -203,11 +207,8 @@ class ViewerMixin:
 
     def show_diff(self):
         """Show the diff for the selected commit"""
-        context = self.context
         self.with_oid(
-            lambda oid: difftool.diff_expression(
-                context, self, oid + '^!', hide_expr=False, focus_tree=True
-            )
+            lambda oid: _diff_expression(self.context, self, oid), filtered=False
         )
 
     def show_dir_diff(self):
@@ -215,8 +216,13 @@ class ViewerMixin:
         context = self.context
         self.with_oid(
             lambda oid: difftool.difftool_launch(
-                context, left=oid, left_take_magic=True, dir_diff=True
-            )
+                context,
+                left=oid,
+                left_take_magic=True,
+                dir_diff=True,
+                staged=oid == dag.STAGE,
+            ),
+            filtered=False,
         )
 
     def rebase_to_commit(self):
@@ -262,12 +268,15 @@ class ViewerMixin:
     def save_blob_dialog(self):
         """Save a file blob from the selected commit"""
         context = self.context
-        self.with_oid(lambda oid: browse.BrowseBranch.browse(context, oid))
+        self.with_oid(
+            lambda oid: browse.BrowseBranch.browse(context, oid), filtered=False
+        )
 
     def save_blob_from_parent_dialog(self):
         """Save a file blob from the parent of the selected commit"""
-        context = self.context
-        self.with_oid(lambda oid: browse.BrowseBranch.browse(context, oid + '^'))
+        self.with_oid(
+            lambda oid: _save_blob_from_parent(self.context, oid), filtered=False
+        )
 
     def update_menu_actions(self, event):
         """Update menu actions to reflect the selection state"""
@@ -279,6 +288,7 @@ class ViewerMixin:
         else:
             self.clicked = commit = item.commit
 
+        has_oid = commit and commit.oid not in (dag.WORKTREE, dag.STAGE)
         has_single_selection = len(selected_items) == 1
         has_single_selection_or_clicked = bool(has_single_selection or commit)
         has_selection = bool(selected_items)
@@ -303,31 +313,56 @@ class ViewerMixin:
         self.menu_actions['diff_selected_this'].setEnabled(can_diff)
         self.menu_actions['diff_commit'].setEnabled(has_single_selection_or_clicked)
         self.menu_actions['diff_commit_all'].setEnabled(has_single_selection_or_clicked)
-
-        self.menu_actions['checkout_branch'].setEnabled(bool(has_branches))
+        self.menu_actions['checkout_branch'].setEnabled(bool(has_branches) and has_oid)
         self.menu_actions['checkout_detached'].setEnabled(
-            has_single_selection_or_clicked
+            has_single_selection_or_clicked and has_oid
         )
-        self.menu_actions['cherry_pick'].setEnabled(has_single_selection_or_clicked)
-        self.menu_actions['copy'].setEnabled(has_single_selection_or_clicked)
-        self.menu_actions['copy_short'].setEnabled(has_single_selection_or_clicked)
-        self.menu_actions['create_branch'].setEnabled(has_single_selection_or_clicked)
-        self.menu_actions['create_patch'].setEnabled(has_selection)
-        self.menu_actions['create_tag'].setEnabled(has_single_selection_or_clicked)
-        self.menu_actions['create_tarball'].setEnabled(has_single_selection_or_clicked)
+        self.menu_actions['cherry_pick'].setEnabled(
+            has_single_selection_or_clicked and has_oid
+        )
+        self.menu_actions['copy'].setEnabled(
+            has_single_selection_or_clicked and has_oid
+        )
+        self.menu_actions['copy_short'].setEnabled(
+            has_single_selection_or_clicked and has_oid
+        )
+        self.menu_actions['create_branch'].setEnabled(
+            has_single_selection_or_clicked and has_oid
+        )
+        self.menu_actions['create_patch'].setEnabled(has_selection and has_oid)
+        self.menu_actions['create_tag'].setEnabled(
+            has_single_selection_or_clicked and has_oid
+        )
+        self.menu_actions['create_tarball'].setEnabled(
+            has_single_selection_or_clicked and has_oid
+        )
         self.menu_actions['rebase_to_commit'].setEnabled(
-            has_single_selection_or_clicked
+            has_single_selection_or_clicked and has_oid
         )
-        self.menu_actions['reset_mixed'].setEnabled(has_single_selection_or_clicked)
-        self.menu_actions['reset_keep'].setEnabled(has_single_selection_or_clicked)
-        self.menu_actions['reset_merge'].setEnabled(has_single_selection_or_clicked)
-        self.menu_actions['reset_soft'].setEnabled(has_single_selection_or_clicked)
-        self.menu_actions['reset_hard'].setEnabled(has_single_selection_or_clicked)
+        self.menu_actions['reset_mixed'].setEnabled(
+            has_single_selection_or_clicked and has_oid
+        )
+        self.menu_actions['reset_keep'].setEnabled(
+            has_single_selection_or_clicked and has_oid
+        )
+        self.menu_actions['reset_merge'].setEnabled(
+            has_single_selection_or_clicked and has_oid
+        )
+        self.menu_actions['reset_soft'].setEnabled(
+            has_single_selection_or_clicked and has_oid
+        )
+        self.menu_actions['reset_hard'].setEnabled(
+            has_single_selection_or_clicked and has_oid
+        )
         self.menu_actions['restore_worktree'].setEnabled(
-            has_single_selection_or_clicked
+            has_single_selection_or_clicked and has_oid
         )
-        self.menu_actions['revert'].setEnabled(has_single_selection_or_clicked)
-        self.menu_actions['save_blob'].setEnabled(has_single_selection_or_clicked)
+        self.menu_actions['revert'].setEnabled(
+            has_single_selection_or_clicked and has_oid
+        )
+        self.menu_actions['save_blob'].setEnabled(
+            has_single_selection_or_clicked and has_oid
+        )
         self.menu_actions['save_blob_from_parent'].setEnabled(
             has_single_selection_or_clicked
         )
@@ -366,6 +401,28 @@ class ViewerMixin:
         menu.addAction(self.menu_actions['copy_short'])
         menu.addAction(self.menu_actions['copy'])
         menu.exec_(self.mapToGlobal(event.pos()))
+
+
+def _diff_expression(context, widget, oid):
+    """Launch difftool using the specified object ID"""
+    if oid == dag.WORKTREE:
+        ref = ''
+    elif oid == dag.STAGE:
+        ref = '--cached'
+    else:
+        ref = f'{oid}^!'
+    return difftool.diff_expression(
+        context, widget, ref, hide_expr=False, focus_tree=True
+    )
+
+
+def _save_blob_from_parent(context, oid):
+    """Save a browse dialog to grab a file from the parent commit"""
+    if oid in (dag.STAGE, dag.WORKTREE):
+        ref = 'HEAD'
+    else:
+        ref = f'{oid}^'
+    return browse.BrowseBranch.browse(context, ref)
 
 
 def set_icon(icon, action):
@@ -956,7 +1013,6 @@ class GitDAG(standard.MainWindow):
     def set_params(self, params):
         context = self.context
         self.params = params
-
         # Update fields affected by model
         self.revtext.setText(params.ref)
         self.maxresults.setValue(params.count)
@@ -964,10 +1020,7 @@ class GitDAG(standard.MainWindow):
 
         if self.thread is not None:
             self.thread.stop()
-
-        self.thread = ReaderThread(context, params, self)
-
-        thread = self.thread
+        self.thread = thread = ReaderThread(context, params, self)
         thread.begin.connect(self.thread_begin, type=Qt.QueuedConnection)
         thread.status.connect(self.thread_status, type=Qt.QueuedConnection)
         thread.add.connect(self.add_commits, type=Qt.QueuedConnection)
@@ -1185,12 +1238,10 @@ class ReaderThread(QtCore.QThread):
 
     def __init__(self, context, params, parent):
         QtCore.QThread.__init__(self, parent)
+        self.setTerminationEnabled(True)
         self.context = context
         self.params = params
-        self._abort = False
         self._stop = False
-        self._mutex = QtCore.QMutex()
-        self._condition = QtCore.QWaitCondition()
 
     def run(self):
         context = self.context
@@ -1199,11 +1250,7 @@ class ReaderThread(QtCore.QThread):
         self.begin.emit()
         commits = []
         for commit in repo.get():
-            self._mutex.lock()
             if self._stop:
-                self._condition.wait(self._mutex)
-            self._mutex.unlock()
-            if self._abort:
                 repo.reset()
                 return
             commits.append(commit)
@@ -1211,30 +1258,24 @@ class ReaderThread(QtCore.QThread):
                 self.add.emit(commits)
                 commits = []
 
-        self.status.emit(repo.returncode == 0)
+        stage, worktree = repo.get_worktree_commits()
+        if stage:
+            commits.append(stage)
+        if worktree:
+            commits.append(worktree)
         if commits:
             self.add.emit(commits)
+
+        self.status.emit(repo.returncode == 0)
         self.end.emit()
 
     def start(self):
-        self._abort = False
         self._stop = False
         QtCore.QThread.start(self)
 
-    def pause(self):
-        self._mutex.lock()
-        self._stop = True
-        self._mutex.unlock()
-
-    def resume(self):
-        self._mutex.lock()
-        self._stop = False
-        self._mutex.unlock()
-        self._condition.wakeOne()
-
     def stop(self):
-        self._abort = True
-        self.wait()
+        self._stop = True
+        self.terminate()
 
 
 class Cache: